Added in ispn base that just checks out the master - note this is cached in the image so you will need to manually pull the latest changes
Also added in pr pull request that does precaching of majority of maven
artifacts to make building faster at the cost of size.  The first
argument to ispn-pr build is for the remote repo and the second is the
branch.  Any additional arguments are passed to mvn clean install if you
want to run only a certain module or etc.